qcacld-3.0: Update band_capability INI to use bitmap
The band_capability field is used in other places to be
a bitmap over reg_wifi_band. Update the field in wma_tgt_cfg and
wlan_mlme_generic to take the same bitmap for consistency.
Update all references to these fields to expect a bitmap as well.

CRs-fixed: 2726710

+++ core/hdd/src/wlan_hdd_wext.c
@@ -3843,7 +3843,7 @@ int wlan_hdd_update_phymode(struct hdd_adapter *adapter, int new_phymode)
 	enum hdd_dot11_mode hdd_dot11mode = hdd_ctx->config->dot11Mode;
 	enum band_info curr_band = BAND_ALL;
 	int retval = 0;
-	uint8_t band_capability;
+	uint32_t band_capability;
 	QDF_STATUS status;
 	uint32_t channel_bonding_mode;
